add error handling for convert script

+++ datasets/convert.sh
@@ -22,5 +22,7 @@ infile="$1"
 if [ -f "$infile" ] ; then
     outfile="$(dirname $1)/$(basename $1 .gz).lzma"
     zcat "$infile" | lzma -9 -c -T 0 --stdout > "$outfile"
-    rm -rf "$infile"
+    if [ $? -eq 0 ]; then
+        rm -rf "$infile"
+    fi
 fi
